Refactored code

+++ src/inference.py
@@ -14,7 +14,6 @@ def run_inference():
     tokenizer = AutoTokenizer.from_pretrained(Config.MODEL_NAME)
     test_loader = get_dataloader(test_df, tokenizer, is_train=False, shuffle=False)
 
-    # Читаємо кращий поріг
     try:
         with open(f"{Config.OUTPUT_DIR}/best_threshold.txt", "r") as f:
             best_th = float(f.read().strip())
@@ -24,11 +23,8 @@ def run_inference():
 
     fold_preds = []
 
-    # Проганяємо через усі 5 фолдів
     for fold in range(Config.N_FOLDS):
-        # Знаходимо файл моделі
         fold_dir = f"{Config.OUTPUT_DIR}/fold_{fold}"
-        # Шукаємо .ckpt файл
         ckpt_files = [f for f in os.listdir(fold_dir) if f.endswith('.ckpt')]
         if not ckpt_files:
             continue
@@ -46,7 +42,7 @@ def run_inference():
                 input_ids = batch['input_ids'].to(Config.DEVICE)
                 attention_mask = batch['attention_mask'].to(Config.DEVICE)
                 logits = model(input_ids, attention_mask)
-                preds = torch.sigmoid(logits).cpu().numpy()
+                preds = torch.sigmoid(logits).float().cpu().numpy()
                 current_preds.extend(preds)
 
         fold_preds.append(np.array(current_preds))
@@ -54,11 +50,9 @@ def run_inference():
         del model
         torch.cuda.empty_cache()
 
-    # Ансамбль (середнє арифметичне)
     avg_preds = np.mean(fold_preds, axis=0)
     final_labels = (avg_preds > best_th).astype(int)
 
-    # Збереження
     submission = pd.read_csv("./input/sample_submission.csv")
     submission['new_label'] = final_labels
     submission.to_csv("submission.csv", index=False)
